feat: moved files and tested some mapping
