+++ src/main/org/fyp24064/model/dto/CreateChatRoomDTO.java
@@ -1,4 +1,4 @@
-package java.org.fyp24064.model.dto;
+package org.fyp24064.model.dto;
 
 import jakarta.persistence.Entity;
 import lombok.Getter;
@@ -6,7 +6,6 @@ import lombok.Setter;
 
 import java.util.List;
 
-@Entity
 @Getter
 @Setter
 public class CreateChatRoomDTO {
