+++ src/main/org/fyp24064/repository/ChatMessageRepository.java
@@ -1,7 +1,7 @@
-package java.org.fyp24064.repository;
+package org.fyp24064.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.org.fyp24064.model.ChatMessage;
+import org.fyp24064.model.ChatMessage;
 
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
