+++ src/test/org/fyp24064/controller/TestPublisherController.java
@@ -1,4 +1,4 @@
-package java.org.fyp24064.controller;
+package org.fyp24064.controller;
 
 public class TestPublisherController {
 }
